refactor(egui): separate module shared
