+++ gui/egui/Cargo.toml
@@ -26,28 +26,20 @@ egui_dnd = "0.14.0"
 egui_extras = "0.33.0"
 # ---
 
-ico = "0.5.0"
-indexmap = { workspace = true, features = ["serde"] }
 mimalloc = { version = "0.1.50", optional = true }
-open = { version = "5.3.5" }
 parking_lot = { workspace = true }
 rapidhash = { workspace = true }
 rayon = { workspace = true }
-semver = { version = "1.0.28", features = ["serde"] }
-serde = { workspace = true }
 snafu = { workspace = true }
-sonic-rs = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
 
 # workspace members
-gh_issue_link = { workspace = true }
 mod_info = { workspace = true }
 nemesis_merge = { workspace = true, features = ["tracing"] }
 skyrim_data_dir = { workspace = true, features = ["serde"] }
 tracing_rotation = { workspace = true }
-
-egui_derive = { path = "../egui_derive" }
+d_merge_gui_shared = { path = "../shared", features = ["tracing"] }
 
 [target.'cfg(not(target_os = "linux"))'.dependencies]
 rfd = { version = "0.17.2", default-features = false }
