+++ gui/egui/src/app/fetch.rs
@@ -27,14 +27,11 @@
 
 use std::sync::Arc;
 
+use d_merge_gui_shared::{fetch::FetchState, i18n::I18nKey, mod_item, settings::DataMode};
 use egui::Color32;
 use rayon::prelude::*;
 
-use crate::{
-    app::{App, state::FetchState},
-    i18n::I18nKey,
-    settings::DataMode,
-};
+use crate::app::App;
 
 impl App {
     /// Triggers an asynchronous refresh of the mod list.
@@ -98,8 +95,7 @@ impl App {
                 drop(state);
 
                 let mod_info = core::mem::take(&mut *self.fetched_mod_info.write());
-                let new_mods =
-                    crate::mod_item::inherit_reorder_cast(self.settings.mod_list(), mod_info);
+                let new_mods = mod_item::inherit_reorder_cast(self.settings.mod_list(), mod_info);
                 self.check_all = new_mods.par_iter().all(|m| m.enabled);
                 *self.settings.mod_list_mut() = new_mods;
