+++ gui/shared/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "d_merge_gui_shared"
+version.workspace = true
+description = "Diff & Merge hkx patcher gui shared code"
+
+authors.workspace = true
+categories = ["development-tools"]
+edition.workspace = true
+keywords = ["skyrim"]
+license = "GPL-3.0"
+readme = "../../README.md"
+repository.workspace = true
+rust-version.workspace = true
+
+
+[dependencies]
+ico = "0.5.0"
+indexmap = { workspace = true, features = ["serde"] }
+open = { version = "5.3.5" }
+parking_lot = { workspace = true }
+rapidhash = { workspace = true }
+rayon = { workspace = true }
+semver = { version = "1.0.28", features = ["serde"] }
+serde = { workspace = true }
+snafu = { workspace = true }
+sonic-rs = { workspace = true }
+tracing = { workspace = true, optional = true }
+
+# workspace members
+gh_issue_link = { workspace = true }
+mod_info = { workspace = true }
+nemesis_merge = { workspace = true, features = ["tracing"] }
+skyrim_data_dir = { workspace = true, features = ["serde"] }
+d_merge_gui_shared_derive = { path = "../shared_derive" }
+
+
+[features]
+tracing = ["dep:tracing"]
+
+
+[lints]
+workspace = true
