+++ gui/shared/src/log.rs
@@ -1,0 +1,42 @@
+//! Start a log file watcher thread.
+//!
+//! This continuously updates `log_lines` with the latest contents of the log file.
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum LogLevel {
+    Error,
+    Warn,
+    Info,
+    Debug,
+    Trace,
+}
+
+impl LogLevel {
+    pub const fn as_str(&self) -> &'static str {
+        match self {
+            Self::Error => "Error",
+            Self::Warn => "Warn",
+            Self::Info => "Info",
+            Self::Debug => "Debug",
+            Self::Trace => "Trace",
+        }
+    }
+}
+
+#[cfg(feature = "tracing")]
+impl From<LogLevel> for tracing::Level {
+    #[inline]
+    fn from(level: LogLevel) -> Self {
+        match level {
+            LogLevel::Error => Self::ERROR,
+            LogLevel::Warn => Self::WARN,
+            LogLevel::Info => Self::INFO,
+            LogLevel::Debug => Self::DEBUG,
+            LogLevel::Trace => Self::TRACE,
+        }
+    }
+}
+
+pub const LOG_DIR: &str = ".d_merge/logs";
+pub const LOG_FILENAME: &str = "d_merge.log";
